Restrict registration to whitelisted email domains.
When REGISTRATION_EMAIL_DOMAIN_WHITELIST configuration variable is present,
users will only be able to register through the registration page if the
domain of their email address is in the list.

There is one exception to this:
Users that were manually added to a course by an instructor will still
be able to register, regardless of what their email address is.