Implement SSO using OpenID.
A single edX instance can serve as the OpenID provider,
while other configured instances will use it as the single sign on endpoint,
and automatically forward users to it when they click the login link.

Forcing single sign on is turned on with the FORCE_OPENID_SSO setting variable.
The OPENID_SSO_SERVER_URL must point to the provider instance.