Fix authentication for OpenID users.
Calling 'login' without first calling 'authenticate' on users doesn't quite work in django.
The 'authenticate' method does something special with the user object to prepare it for login.

Appearantly it can be worked around - and has been worked around in the code for shibboleth users.
Do the same for users coming from the SSO OpenID provider.