Test if JSINPUT tests are really running
This test currently fails, meaning that the existing tests weren't
testing what they claimed.

Assertions are made for each element returned by the CSS selectors.
However, the selectors are assumed to be wildcard matches, but are
actually literal selectors. As there are no matched elements, this
causes the assertions to be (silently) checked zero times, without
failure.